+++ game_engine/tic_tac_toe_common_lib.py
@@ -1,4 +1,4 @@
-from abc import ABC, abstractmethod
+from abc import ABC, abstractmethod, abstractproperty
 from typing import List, Callable
 from dataclasses import dataclass
 
@@ -18,6 +18,12 @@ class TicTacToeGameInfo:
     first_player_id: str
     second_player_id: str
     winner_id: str # а какие могут быть варианты?
+
+
+@dataclass
+class UserInfo:
+    user_id: str
+    secret_key: str
 
 
 class AbstractTicTacToeGame(ABC):
@@ -43,4 +49,3 @@ class AbstractTicTacToeGame(ABC):
     @abstractmethod
     def get_game_info(self) -> TicTacToeGameInfo:
         """обычный геттер"""
-
